Create camper leaderboard app
Added more styling to the table to make more appealing to end user. There is a checkmark to indicate which column the table is sorted by. Hover added on table row to help emphasize data for user. Refactored the render function and removed the code that builds the table rows to aseperate function.